Add demo to retrieve image URI
Added a non-permanent demo button to demonstrate the execution of the
tokenURI function that is described in the abi.  Note that the abi file
contains functions from inherited contracts.

The URI will be key to constructing the URL to pinata for the display of
hte files.  The URI on the item is set in the createToken function of
nft.sol.  We'll need to consider our format for the URI so that it is
descriptive and timeless, but also easily able to convert to a URL.
:wq
https://gateway.pinata.cloud/ipfs/